requirements.txt
python-binance==1.0.28
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
websocket-client==1.8.0